Move to me.hugmanrique.cellarium package
