+++ src/main/java/me/hugmanrique/cellarium/ForwardingRepository.java
@@ -1,67 +1,80 @@
 package me.hugmanrique.cellarium;
 
+import javax.annotation.Nullable;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.function.BiFunction;
-import java.util.function.UnaryOperator;
 
 /**
  * A repository that forwards all its method calls to another repository.
- * Subclasses can override any method to modify the behavior of the
+ * Subclasses can override any method to modify the default behavior of the
  * backing repository.
  *
  * @see <a href="https://en.wikipedia.org/wiki/Decorator_pattern">Decorator pattern</a>
  */
 public class ForwardingRepository implements Repository {
+
     private final Repository repository;
 
     public ForwardingRepository(Repository repository) {
-        this.repository = Objects.requireNonNull(repository);
+        this.repository = Objects.requireNonNull(repository, "repository");
+    }
+
+    @Nullable
+    @Override
+    public <T> T get(Key<T> key) {
+        return repository.get(key);
+    }
+
+    @Nullable
+    @Override
+    public <T> T put(Key<T> key, T value) {
+        return repository.put(key, value);
+    }
+
+    @Nullable
+    @Override
+    public <T> T putIfAbsent(Key<T> key, T value) {
+        return repository.putIfAbsent(key, value);
+    }
+
+    @Nullable
+    @Override
+    public <T> T replace(Key<T> key, T value) {
+        return repository.replace(key, value);
     }
 
     @Override
-    public boolean hasValue(Item<?> item) {
-        return repository.hasValue(item);
+    public <T> boolean replace(Key<T> key, T oldValue, T newValue) {
+        return repository.replace(key, oldValue, newValue);
+    }
+
+    @Nullable
+    @Override
+    public <T> T remove(Key<T> key) {
+        return repository.remove(key);
     }
 
     @Override
-    public <T> Optional<T> getValue(Item<T> item) {
-        return repository.getValue(item);
+    public <T> T remove(Key<T> key, T value) {
+        return repository.remove(key, value);
     }
 
     @Override
-    public <T> Optional<T> setValue(Item<T> item, T value) {
-        return repository.setValue(item, value);
+    public void clear() {
+        repository.clear();
     }
 
     @Override
-    public <T> Optional<T> apply(Item<T> item, UnaryOperator<T> operation) {
-        return repository.apply(item, operation);
+    public boolean contains(Key<?> key) {
+        return repository.contains(key);
     }
 
     @Override
-    public <T> Optional<T> apply(Item<T> item, BiFunction<Item<T>, T, T> operation) {
-        return repository.apply(item, operation);
+    public int size() {
+        return repository.size();
     }
 
     @Override
-    public <T> Optional<T> removeValue(Item<T> item) {
-        return repository.removeValue(item);
-    }
-
-    @Override
-    @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
-    public boolean equals(Object obj) {
-        return repository.equals(obj);
-    }
-
-    @Override
-    public int hashCode() {
-        return repository.hashCode();
-    }
-
-    @Override
-    public String toString() {
-        return repository.toString();
+    public boolean isEmpty() {
+        return repository.isEmpty();
     }
 }
