+++ src/main/java/me/hugmanrique/cellarium/Key.java
@@ -1,4 +1,4 @@
-package me.hugmanrique.cellarium.v2;
+package me.hugmanrique.cellarium;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
