+++ src/main/java/me/hugmanrique/cellarium/oldutil/EnumItems.java
@@ -1,7 +1,4 @@
-package me.hugmanrique.cellarium.util;
-
-import me.hugmanrique.cellarium.Item;
-import me.hugmanrique.cellarium.Repository;
+package me.hugmanrique.cellarium.oldutil;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
