+++ src/main/java/me/hugmanrique/cellarium/simple/package-info.java
@@ -1,4 +1,4 @@
 @ParametersAreNonnullByDefault
-package me.hugmanrique.cellarium.v2.simple;
+package me.hugmanrique.cellarium.simple;
 
 import javax.annotation.ParametersAreNonnullByDefault;
